# Try and force precompilation