@chriselrod 2021-08-07
Is 0.04 seconds worth it?
How much longer does precipitation take?