@ChrisRackauckas 2021-08-07
not worth it to merge now. But keeping this up to refer back to.