@ChrisRackauckas 2021-08-07
It's 0.3 seconds BTW.