@chriselrod 2021-08-09
> not worth it to merge now. But keeping this up to refer back to.

This seems to substantially improve time to-first-div, on this PR:
```julia
julia> @time using TriangularSolve, LinearAlgebra
[ Info: Precompiling TriangularSolve [d5829a12-d9aa-46ab-831f-fb7c9ab06edf]
 24.632061 seconds (8.20 M allocations: 456.805 MiB, 0.58% gc time, 0.09% compilation time)

julia> A = rand(1,1);

julia> B = rand(1, 1);

julia> res = similar(A);

julia> @t TriangularSolve.rdiv!(res, A, UpperTriangular(B))
Time = 0.901624834
1×1 Matrix{Float64}:
 0.9805494758920018

julia> @t TriangularSolve.rdiv!(res, A, UnitUpperTriangular(B))
Time = 0.08022014000000001
1×1 Matrix{Float64}:
 0.04453566915648233

julia> @t TriangularSolve.ldiv!(res, LowerTriangular(B), A)
Time = 0.195168043
1×1 Matrix{Float64}:
 0.9805494758920018

julia> @t TriangularSolve.ldiv!(res, UnitLowerTriangular(B), A)
Time = 0.094892007
1×1 Matrix{Float64}:
 0.04453566915648233
```
Main:
```julia
julia> @time using TriangularSolve, LinearAlgebra
[ Info: Precompiling TriangularSolve [d5829a12-d9aa-46ab-831f-fb7c9ab06edf]
  6.625835 seconds (7.92 M allocations: 430.333 MiB, 1.97% gc time, 0.33% compilation time)

julia> A = rand(1,1);

julia> B = rand(1, 1);

julia> res = similar(A);

julia> @t TriangularSolve.rdiv!(res, A, UpperTriangular(B))
Time = 16.214664716
1×1 Matrix{Float64}:
 55.090822966245106

julia> @t TriangularSolve.rdiv!(res, A, UnitUpperTriangular(B))
Time = 0.337600035
1×1 Matrix{Float64}:
 0.9878212675272343

julia> @t TriangularSolve.ldiv!(res, LowerTriangular(B), A)
Time = 8.620501925000001
1×1 Matrix{Float64}:
 55.090822966245106

julia> @t TriangularSolve.ldiv!(res, UnitLowerTriangular(B), A)
Time = 0.563920552
1×1 Matrix{Float64}:
 0.9878212675272343
```