@timholy 2021-08-09
![image](https://user-images.githubusercontent.com/1525481/128686250-811acfa7-c0fa-4d45-9c43-fa5c537abb2f.png)

The red bars are not precompilable, but the other colors are. The right place to do this precompilation is DiffEqBase; if you click on the bar where my mouse pointer is, in the REPL you'll see something like this:
```julia
/tmp/pkgs/packages/DiffEqBase/uQlhE/src/linear_nonlinear.jl:93, MethodInstance for (::DefaultLinSolve)(::Vector{Float64}, ::Any, ::Vector{Float64}, ::Bool)
```
That gives you an indication of what needs to be precompiled. You could use SnoopCompile's `parcel` and `write` to generate these, but as we discussed previously these days I strongly recommend just running some code while the package is building, unless you can't avoid the undesirable side effects.