@ranocha 2021-08-09
> You could use SnoopCompile's `parcel` and `write` to generate these, but as we discussed previously these days I strongly recommend just running some code while the package is building, unless you can't avoid the undesirable side effects.

I didn't follow that discussion - do you have some hints or references for further information?