@timholy 2021-08-09
@ranocha, see https://timholy.github.io/SnoopCompile.jl/stable/snoopi_deep_parcel/. Everybody *rushes* to pick up this hammer, but it's literally the last thing you should do when trying to reduce latency. The right steps are:
1. check that you're not getting a lot of invalidations (fix them if you are)
2. look at the inference profile flamegraph(like the image above) and see if anything pops out (in this case, it does, but that's relatively rare and usually means that the package is in great shape except possibly for missing a bit of precompilation)
3. use `pgdsgui` to determine whether you are over-specializing, and adjust as needed
4. solve inference problems (might need to look at the previous step again a bit)
5. see if you can precompile by doing "regular work" rather than issuing precompile directives (looking at the flamegraph again here is very helpful; after you fix inference problems, see the previous step, hopefully you don't have very many "flames" to check just like in this case)
6. as a last resort, use `parcel` and `write`

I almost actively dislike `parcel` and `write` as they often hide problems that are better addressed by other means, and often gives you very little advantage compared to a more thorough analysis (e.g., https://github.com/JuliaIO/JLD2.jl/pull/344). As both SnoopCompile and JET have discovered, snooping on inference is a terrific way to discover things that many packages should probably change anyway.