@ChrisRackauckas 2021-08-12
Downstream DiffEq had to end up precompile this call differently, but most users will get the benefit from this PR.