@chriselrod 2021-08-29
This PR caused a performance regression for me.
With precompilation:
```julia
julia> @benchmark TriangularSolve.rdiv!($C, $A, UpperTriangular($B), Val(false)) # false means single threaded
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  18.276 μs …  1.964 ms  ┊ GC (min … max): 0.00% … 97.10%
 Time  (median):     21.517 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   22.028 μs ± 19.593 μs  ┊ GC (mean ± σ):  0.87% ±  0.97%

             █▁▁▆▃ ▂ ▁
  ▃▃▃▃▃▃▃▃▃▇▄█████▇████▇▆▅▄▃▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ ▃
  18.3 μs         Histogram: frequency by time        30.7 μs <

 Memory estimate: 6.39 KiB, allocs estimate: 211.
```
Without:
```julia
julia> @benchmark TriangularSolve.rdiv!($C, $A, UpperTriangular($B), Val(false)) # false means single threaded
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  15.746 μs … 44.410 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     18.713 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   18.684 μs ±  1.184 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▂▁                 ▁▂         ▂▅▃   ▅█▆    ▁▂     ▁         ▁
  ██▃▁▆█▆▃▁▁▁▄▁▃▄▄▄▁▁██▇▃▃▁▃▁▅▃▅███▅▃▄████▅▅▆██▇▅▅▆███▇▇▆▆▆█▇ █
  15.7 μs      Histogram: log(frequency) by time      20.4 μs <

 Memory estimate: 0 bytes, allocs estimate: 0.
```
Notice the memory allocations in particular. =/