@chriselrod 2021-08-29
> [SciML/OrdinaryDiffEq.jl#1473](https://github.com/SciML/OrdinaryDiffEq.jl/pull/1473)

Hmm, okay, I'll try playing around with type signatures.

I noticed that adding a few `@inline`s could reduce the number of allocations to 3-6.