@chriselrod 2021-08-29
The weird thing is that the functions it helps to inline are `@generated`, which I thought are always supposed to fully specialize.

I take this to mean that the function calling them isn't specializing?